dailybuild: add/modify VS2015 test solutions (src/Samples)
git-svn-id: http://springhead.info/spr2/Springhead2/trunk@8428 05cee5c3-a2e9-0310-9523-9dfc2f93dbe1

+++ src/Samples/Foundation/UTTimerSample/main.cpp
@@ -18,7 +18,7 @@ class TimerProviderGL: public UTTimerProvider{
 	typedef std::vector<Arg*> Args;
 	Args args;
 	static void SPR_CDECL Callback(int a){
-		Arg* arg = (Arg*)a;
+		Arg* arg = (Arg*)(size_t)a;
 		if (!arg->bStop){
 			arg->timer->Call();
 			glutTimerFunc(arg->timer->GetInterval(), Callback, a);
@@ -35,7 +35,7 @@ public:
 	}
 	bool StartTimer(UTTimer* timer){
 		args.push_back(new Arg(timer));
-		glutTimerFunc(timer->GetInterval(), Callback, (int)(void*)args.back());
+		glutTimerFunc(timer->GetInterval(), Callback, (int)(size_t)(void*)args.back());
 		return true;
 	}
 	bool StopTimer(UTTimer* timer){
